pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "alembic-postgresql-enum>=1.8.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.125.0",
    "greenlet>=3.3.0",
    "humps>=0.2.2",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
